fix init order & add cli args
